+++ R/load_model.R
@@ -1,6 +1,6 @@
 #' @include Ensemble.SDM.R Stacked.SDM.R
 #' @importFrom shiny incProgress
-#' @importFrom raster raster stack
+#' @importFrom raster raster stack reclassify
 NULL
 
 #' Function to load ensemble SDMs and SSDMs
@@ -32,15 +32,22 @@ load_enm = function (name, path = getwd()) {
     cat('Algorithm correlation table empty ! \n')
     a = data.frame()
   }
+  b = try(raster(paste0(path,'/Rasters/Probability.tif')))
+  if (inherits(b, 'try-error')) {
+    projection = raster(paste0(path,'/Rasters/Probability.tif'))
+    evaluation = read.csv(paste0(path,'/Tables/ENMeval.csv'), row.names = 1)
+    b = reclassify(projection, c(-Inf,evaluation$threshold,0, evaluation$threshold,Inf,1))
+  }
   enm = Ensemble.SDM(name = as.character(read.csv(paste0(path,'/Tables/Name.csv'))[1,2]),
-                             projection = raster(paste0(path,'/Rasters/Probability.tif')),
-                             uncertainty = try(raster(paste0(path,'/Rasters/uncertainty.tif'))),
-                             evaluation = read.csv(paste0(path,'/Tables/ENMeval.csv'), row.names = 1),
-                             algorithm.evaluation  = read.csv(paste0(path,'/Tables/AlgoEval.csv'), row.names = 1),
-                             algorithm.correlation = a,
-                             data = read.csv(paste0(path,'/Tables/Data.csv'), row.names = 1),
-                             variable.importance = read.csv(paste0(path,'/Tables/VarImp.csv'), row.names = 1),
-                             parameters = read.csv(paste0(path,'/Tables/Parameters.csv'), row.names = 1, colClasses = "character"))
+                     projection = raster(paste0(path,'/Rasters/Probability.tif')),
+                     # Add binary if file open it if not create it for compatibility
+                     uncertainty = try(raster(paste0(path,'/Rasters/uncertainty.tif'))),
+                     evaluation = read.csv(paste0(path,'/Tables/ENMeval.csv'), row.names = 1),
+                     algorithm.evaluation  = read.csv(paste0(path,'/Tables/AlgoEval.csv'), row.names = 1),
+                     algorithm.correlation = a,
+                     data = read.csv(paste0(path,'/Tables/Data.csv'), row.names = 1),
+                     variable.importance = read.csv(paste0(path,'/Tables/VarImp.csv'), row.names = 1),
+                     parameters = read.csv(paste0(path,'/Tables/Parameters.csv'), row.names = 1, colClasses = "character"))
   return(enm)
 }