+++ R/load_occ.R
@@ -1,6 +1,6 @@
 #' @include checkargs.R
 #' @importFrom shiny incProgress
-# @importFrom spThin thin
+#' @importFrom spThin thin
 #' @importFrom raster res extract
 NULL